requirements.txt
streamlit
pandas
python-dotenv
supabase
xlrd
openpyxl
